requirements.txt
asgiref==3.7.2
Django==5.0
django-dotenv==1.4.2
djangorestframework==3.14.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
